+++ operations-api/build.gradle.kts
@@ -27,7 +27,11 @@ kotlin {
         }
     }
     sourceSets {
-        val commonMain by getting
+        val commonMain by getting {
+            dependencies {
+                implementation(project(Modules.utils))
+            }
+        }
         val commonTest by getting {
             dependencies {
                 implementation(kotlin("test"))
