+++ operations-stdlib/src/commonMain/kotlin/Find.kt
@@ -1,12 +1,12 @@
 import unwrap.EvaluatingUnwrapper
-import unwrap.getMappingOperationOrNull
 import utils.asList
+import utils.getMappingOperationOrNull
 
 object Find : FunctionalLogicOperation, EvaluatingUnwrapper {
     override fun invoke(expression: Any?, data: Any?, evaluator: LogicEvaluator): Any? {
         return expression.asList.let { expressionValues ->
             val inputData = unwrapDataByEvaluation(expressionValues, data, evaluator)
-            val predicateOperation = getMappingOperationOrNull(expressionValues)
+            val predicateOperation = expressionValues.getMappingOperationOrNull()
 
             predicateOperation?.let {
                 inputData?.find { evaluator.evaluateLogic(predicateOperation, it) == true }
