+++ utils/build.gradle.kts
@@ -2,7 +2,8 @@ import org.jetbrains.kotlin.gradle.plugin.mpp.apple.XCFramework
 
 plugins {
     kotlin("multiplatform") version Versions.kotlin
-    id("io.kotest.multiplatform") version Versions.kotest apply false
+    id("io.kotest.multiplatform") version Versions.kotest
+    id(Conventions.junit)
 }
 
 kotlin {
@@ -26,8 +27,19 @@ kotlin {
     }
     sourceSets {
         val commonMain by getting
-        val commonTest by getting
-        val jvmTest by getting
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test"))
+                implementation(Libs.Kotest.assertionsCore)
+                implementation(Libs.Kotest.frameworkEngine)
+                implementation(Libs.Kotest.frameworkDataset)
+            }
+        }
+        val jvmTest by getting {
+            dependencies {
+                implementation(Libs.Kotest.jvmJunit5Runner)
+            }
+        }
         val iosX64Main by getting
         val iosArm64Main by getting
         val iosSimulatorArm64Main by getting
